index variable in OpenMP 'for' statement must have signed integral type (msvc)
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@1757 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ features/include/pcl/features/impl/fpfh_omp.hpp
@@ -82,7 +82,7 @@ pcl::FPFHEstimationOMP<PointInT, PointNT, PointOutT>::computeFeature (PointCloud
   // Compute SPFH signatures for every point that needs them
   
 #pragma omp parallel for schedule (dynamic, threads_)
-  for (size_t i = 0; i < spfh_indices_vec.size (); ++i)
+  for (int i = 0; i < (int) spfh_indices_vec.size (); ++i)
   {
     // Get the next point index
     int p_idx = spfh_indices_vec[i];
@@ -104,7 +104,7 @@ pcl::FPFHEstimationOMP<PointInT, PointNT, PointOutT>::computeFeature (PointCloud
 
   // Iterate over the entire index vector
 #pragma omp parallel for schedule (dynamic, threads_)
-  for (size_t idx = 0; idx < indices_->size (); ++idx)
+  for (int idx = 0; idx < (int) indices_->size (); ++idx)
   {
     // Find the indices of point idx's neighbors...
     std::vector<int> nn_indices (k_); // \note These resizes are irrelevant for a radiusSearch ().
